% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{aggregate_data_from_config}
\alias{aggregate_data_from_config}
\title{Aggregate a given data.table object based on config}
\usage{
aggregate_data_from_config(df, config, ...)
}
\arguments{
\item{df}{data.table::data.table - Object containing the data to be aggregated}

\item{config}{list - Configuration list from the config file}

\item{...}{\itemize{
\item Any number of additional named arguments to pass through to the aggregator and in turn to collapse::collap()
}}
}
\value{
Optional\url{data.table::data.table|Error}
}
\description{
Wrapper around the aggregator function to aggregate a given data set with the specified configuration block
within the configuration file.
}
