% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{aggregator}
\alias{aggregator}
\title{Agggregate target input data}
\usage{
aggregator(
  df,
  index_vars = NULL,
  numerical_vars = NULL,
  categorical_vars = NULL,
  agg_methods_num = NULL,
  agg_methods_cat = NULL,
  weight_vars = NULL,
  use_weight = FALSE,
  ...
)
}
\arguments{
\item{df}{data.table::data.table - An object of type data.frame}

\item{index_vars}{vec\if{html}{\out{<character>}} - A column name or vector of column names present in \code{data} to aggregate by}

\item{numerical_vars}{vec\if{html}{\out{<character>}} - A vector of column names for numerical aggregations}

\item{categorical_vars}{vec\if{html}{\out{<character>}} - A vector of column names for categorical aggregations}

\item{agg_methods_num}{vec\if{html}{\out{<character>}} - A vector of aggregation methods for each specified column in \code{numerical_vars}}

\item{weight_vars}{vec\if{html}{\out{<character>}} - One or more present columns in the passed \code{df} which should be used for weighting}

\item{use_weight}{A logical vector with length \code{length(c(numerical_vars, categorical_vars))} or a single
\code{TRUE} / \code{FALSE} for global option setting if weights should be used}

\item{...}{Additional parameters which can be passed to \code{collapse::collap()}}

\item{agg_methods_categorical}{vec\if{html}{\out{<character>}} - A vector of aggregation methods for each specified column in \code{categorical_vars}}
}
\description{
This function takes at least one index column as aggregation level as well as at least on numerical or categorical
variable as input arguments with their respective aggregation functions. The aggregation method for categorical
values defaults to their mode, e.g. the most frequent value. The aggregator is capable of advanced weighted and
grouped custom aggregations by construction a call to the \code{custom} argument from \code{collapse::collap()}.
Possible arguments are existing columns within the given data.frame for \code{index_vars},
\code{numerical_columns}, \code{categorical_columns} and \code{weight_vars}. Multiple weight columns can be
chosen. Possible aggregation methods for numerical columns are \code{fsum}, \code{fmean}, \code{fmax},
\code{fmin}, \code{fmedian}, \code{fsd}, \code{fvar} or \code{flast}. All mentioned columns originate in the
\code{collapse} R package for fast statistical programming. With the optional parameter \code{use_weight} it can be
specified if weights should apply and if yes, to which columns. Default is \code{NULL} for no weighted aggregations.
The parameter can handle a vector of length \code{length(c(numerical_columns, categorical_columns))} if you wish to
perform weighted aggregations with the chosen method only on a subset of the passed data.
}
