% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{all_false}
\alias{all_false}
\title{Check if all are FALSE}
\usage{
all_false(...)
}
\arguments{
\item{...}{\itemize{
\item Any number of unnamed input arguments
}}
}
\value{
bool
}
\description{
Small helper function to check if all of the provided input arguments are FALSE.
}
