% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{any_null}
\alias{any_null}
\title{Check if any is NULL}
\usage{
any_null(...)
}
\arguments{
\item{...}{\itemize{
\item Any number of unnamed input arguments
}}
}
\value{
bool
}
\description{
Small helper function to check if any of the provided input arguments is NULL.
}
