% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_fitnessfunc.R
\name{calc_fitness_score}
\alias{calc_fitness_score}
\title{Calculate fitness score}
\usage{
calc_fitness_score(
  individual,
  df,
  categorical_vars = NULL,
  numerical_vars = NULL
)
}
\arguments{
\item{individual}{vec\if{html}{\out{<Some>}} - Vector containing the genome representation of the individual}

\item{df}{data.table::data.table - Object containing all the (prepared and aggregated) data of the individual}

\item{numerical_vars}{vec\if{html}{\out{<character>}} - Vector containing all column names of numerical columns for the calculation}

\item{categorical_columns}{vec\if{html}{\out{<character>}} - Vector containing all column names of categorical columns for the calculation}
}
\value{
numeric - Fitness score value of the individual
}
\description{
TODO DESCRIPTION
}
