% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_fitnessfunc.R
\name{calculate_distance}
\alias{calculate_distance}
\title{Distance function}
\usage{
calculate_distance(X)
}
\arguments{
\item{X}{matrix - A matrix containing the distributions of a group per row (!)}
}
\value{
numeric - Exact (if nrows(X) = 2) or weighted mean distance between all passed groups
}
\description{
This function calculates the distance between n groups within a given individual and its given attributes,
e.g. data values from the specified numerical and/or categorical variables within the configuration.
If there are only two groups we rely on philentropy::euclidean to do the calculations, else it is run via
philentropy::distance obtaining a distance object. This distance object, possibly containing very large difference
between some and smaller distances between other groups, is condensed into a single "mean distance" value,
calculated as the Root Mean Squared (RMS) Distance.
}
