% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{categorical_aggregation}
\alias{categorical_aggregation}
\title{Categorical aggregator}
\usage{
categorical_aggregation(df, cols, method = "freq")
}
\arguments{
\item{df}{A data.table object}

\item{cols}{The categorical columns for frequency caluclations}

\item{method}{The method passed to \code{frequency_values}}
}
\value{
A data.table with the method result and input columns specified
}
\description{
This function takes a data.table object and a vector of categorical columns as input and returns the specified mehod
result of each unique column entry relative to the total amount of data entries passed.
}
