% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ab_division_data.R
\name{create_ab_division_data}
\alias{create_ab_division_data}
\title{Take transactions and items to create final data}
\usage{
create_ab_division_data(items, transactions, items_config, transactions_config)
}
\arguments{
\item{items}{data.table::data.table - Object containing the items data}

\item{transactions}{data.table::data.table - Object containing the transactions data}

\item{items_config}{list - Configuration list}

\item{transactions_config}{list - Configuration list}
}
\value{
Optional\url{data.table::data.table|error}
}
\description{
Given the items and transactions data as well as their respective config blocks,
this function creates the final data set to use within the AB Division algorithm
chosen.
}
