% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{create_formula}
\alias{create_formula}
\title{Create a formula object}
\usage{
create_formula(right_side, left_side = NULL)
}
\arguments{
\item{right_side}{vec\if{html}{\out{<character>}}}

\item{left_side}{vec\if{html}{\out{<character>}}}
}
\value{
formula
}
\description{
Helper function to create a formula object based on one or two character vectors
}
