% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_mutation.R
\name{flipping_mutation}
\alias{flipping_mutation}
\title{Flipping Mutation}
\usage{
flipping_mutation(individual, mutation_probability = 0.5, seed = NULL)
}
\arguments{
\item{individual}{vec\if{html}{\out{<bool>}} - A boolean vector representating the genome information}

\item{mutation_probability}{numeric - The probability of a mutation happening}

\item{seed}{Optional<NULL|integer> - A random seed for deterministic outcomes}
}
\value{
Optional<err|vec\if{html}{\out{<bool>}}>
}
\description{
Flipping mutation is a mutation only for binary encoded individuals. Given the mutation_probability and individual
it iterates through the boolean vector of individual and generates a random number between 0.0 and 1.0. If the
number is less than mutation_probability then we flip the value at that index else it remains the same.
The function can also take in an optional seed value for deterministic results.
}
