% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{frequency_values}
\alias{frequency_values}
\title{Frequency Values function}
\usage{
frequency_values(vec, method = NULL)
}
\arguments{
\item{vec}{vec\if{html}{\out{<character>}} - Vector of categorical data for aggregation}

\item{method}{character - Character string specifying the method to use. One of "count", "freq" or "inv_freq"}
}
\value{
data.table::data.table - Object containing aggregated and to numerical values transformed data
}
\description{
Frequency Values function
}
