% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{get_aggregation_methods}
\alias{get_aggregation_methods}
\title{Helper function to parse methods from config}
\usage{
get_aggregation_methods(config, cols)
}
\arguments{
\item{config}{A configuration list}

\item{cols}{A vector of column names}
}
\value{
Optional<vec\if{html}{\out{<character>}}, NULL>
}
\description{
Helper function to parse methods from config
}
