% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{get_aggregation_weights}
\alias{get_aggregation_weights}
\title{Helper function to find the columns specified as weights}
\usage{
get_aggregation_weights(config, cols)
}
\arguments{
\item{config}{A configuration list}

\item{cols}{A vector of column names}
}
\value{
Optional<vec\if{html}{\out{<character>}},NULL>
}
\description{
Helper function to find the columns specified as weights
}
