% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{get_categorical_vars}
\alias{get_categorical_vars}
\title{Parse all categorical attributes from a given configuration}
\usage{
get_categorical_vars(config)
}
\arguments{
\item{config}{A configuration list}
}
\description{
Function to extract, parse and return all categorical variables within a
provided configuration. See exmaple configuration.
}
