% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{get_numerical_vars}
\alias{get_numerical_vars}
\title{Parse all numerical variables from a given configuration}
\usage{
get_numerical_vars(config)
}
\arguments{
\item{config}{A configuration list}
}
\value{
list<vec<character|bool>>
}
\description{
Function to extract, parse and return all numerical variables within a
provided configuration. See example configuration.
Returns a list of vectors for numerical columns, weights, methods as well
as the weight column.
}
