% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{get_strategy_config}
\alias{get_strategy_config}
\title{Helper function to receive the algorithm config}
\usage{
get_strategy_config(config)
}
\arguments{
\item{config}{list - A configuration list, see example configuration}
}
\value{
list
}
\description{
Helper function to receive the algorithm config
}
