% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{get_use_of_weights}
\alias{get_use_of_weights}
\title{Helper function to parse weights from config}
\usage{
get_use_of_weights(config, cols)
}
\arguments{
\item{config}{A configuration list}

\item{cols}{A vector of column names}
}
\value{
Optional<vec\if{html}{\out{<bool>}},NULL>
}
\description{
This functions extracts the information about the use of a weight
from the given configuration for all numerical columns
}
