% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{get_vars_by_type}
\alias{get_vars_by_type}
\title{Helper function to receive columns from config by type}
\usage{
get_vars_by_type(config, types = NULL)
}
\arguments{
\item{config:}{list A configuration list}

\item{types:}{vec\if{html}{\out{<character>}} A vector of types to get variables from}
}
\value{
Optional<vec\if{html}{\out{<character>}}, NULL>
}
\description{
Helper function to receive columns from config by type
}
