% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_mutation.R
\name{inverse_mutation}
\alias{inverse_mutation}
\title{Inverse Mutation}
\usage{
inverse_mutation(individual, seed = NULL)
}
\arguments{
\item{individual}{vec\if{html}{\out{<bool>}} - A boolean vector representating the genome information}

\item{seed}{Optional<NULL|integer> - A random seed for deterministic outcomes}
}
\value{
Optional<err|vec\if{html}{\out{<bool>}}>
}
\description{
Inversion mutation is a mutation only for binary encoded individuals. Given the individual it randomly generates
two indices and then inverts the value between those indices of the individual. The function can also take in an
optional seed value for deterministic results.
}
