% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Get target data based on a given configuration block}
\usage{
load_data(config = NULL)
}
\arguments{
\item{config}{list: Parsed YAML configuration from config::get() as list}
}
\value{
data.table::data.table
}
\description{
Requires a configuration list for a data input. See example configuration.
}
