% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliaries.R
\name{mutate_genes}
\alias{mutate_genes}
\title{Mutate Genes Function}
\usage{
mutate_genes(len, valid_genes)
}
\arguments{
\item{len}{integer - An integer specifying the number of genes to sample}

\item{valid_genes}{vec<Optional<character|numeric>> - A vector of valid genes from which to sample}
}
\value{
A vector of randomly sampled genes.
}
\description{
This function generates a random sample of genes from a given set of valid genes.
}
