% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_session.R
\name{parse_cli_options}
\alias{parse_cli_options}
\title{Parse CLI options}
\usage{
parse_cli_options()
}
\value{
list - List of parsed CLI options
}
\description{
This function is in charge of finding as well as processing all given
command line arguments. They are defined within this function for now
as well, to keep it simple. This might change in future versions.
}
