% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_config.R
\name{parse_data_config}
\alias{parse_data_config}
\title{Parse a configuration part from items or translogs}
\usage{
parse_data_config(config)
}
\arguments{
\item{config}{A configuration object}
}
\description{
Function to parse a given config received via config::get() with the
necessary structure provided by the example configuration. Returns a named
list of all valid arguments needed for the genetic algorithm funcitonality
}
