% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_session.R
\name{prepare_ab_division_session}
\alias{prepare_ab_division_session}
\title{Prepare A/B Division Session}
\usage{
prepare_ab_division_session()
}
\value{
A configuration list for the further use in the program
}
\description{
Wrapper around some utility functions to set up, load and check all
requirements of the A/B Division program. Should be run within every
run to make sure the provided configuration and options are valid.
}
