% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_fitnessfunc.R
\name{prepare_distance_calculation}
\alias{prepare_distance_calculation}
\title{Prepare distance calculation}
\usage{
prepare_distance_calculation(
  df,
  index_column,
  categorical_vars = NULL,
  numerical_vars = NULL
)
}
\arguments{
\item{df}{data.table::data.table - Object containing the numerical and categorical data of the individual}

\item{index_column}{character - Column name of the index column}

\item{categorical_vars}{vec\if{html}{\out{<character>}} - Character vector of column names indicating categorical variables}

\item{numerical_vars}{vec\if{html}{\out{<character>}} - Character vector of column names indicating numerical variables}
}
\description{
This function is responsible of preparing the given data w.r.t. the group division within a given individual.
}
