% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_session.R
\name{prepare_runtime_session}
\alias{prepare_runtime_session}
\title{Prepare A/B Division runtime session}
\usage{
prepare_runtime_session()
}
\description{
This function provides the basic setup functionalities to perorm the A/B
division algorithms. This includes checking accessibility of available
cores for parallel processing.
}
