% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_mutation.R
\name{random_mutation}
\alias{random_mutation}
\title{Random Mutation}
\usage{
random_mutation(
  individual,
  valid_genes = c(TRUE, FALSE),
  mutation_probability = 0.5,
  seed = NULL
)
}
\arguments{
\item{individual}{vec\if{html}{\out{<bool>}} - A boolean vector representating the genome information}

\item{valid_genes}{vec\if{html}{\out{<Some>}} - Vector of possible values for a gene}

\item{mutation_probability}{numeric - The probability of a mutation happening}

\item{seed}{Optional<NULL|integer> - A random seed for deterministic outcomes}
}
\value{
Optional<err|vec\if{html}{\out{<bool>}}>
}
\description{
Random mutation is for a valid gene set larger than two. Given the set of valid genes as well as a mutation
probability this function mutates a specific genome if the drawn probability exceeds the provided mutation
probability. with any from the given valid genes except the current one present. The function can also take in an
optional seed value for deterministic results.
}
