% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_selection.R
\name{random_selection}
\alias{random_selection}
\title{Random selection}
\usage{
random_selection(fitness_values, num_parents, seed = NULL)
}
\arguments{
\item{fitness_values}{vec\if{html}{\out{<numeric>}} - A vector of numeric values representing the individuals fitness values}

\item{num_parents}{integer - The number of individuals to choose}

\item{seed}{Optional<NULL|integer> - A random seed for deterministic outcomes}
}
\value{
vec\if{html}{\out{<integer>}}
}
\description{
This function selects a random set of individuals from a given
vector of fitness scores. The function optionally accepts a seed for deterministic outcomes.
}
