% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_session.R
\name{setup_environment_config}
\alias{setup_environment_config}
\title{Setup environment configuration}
\usage{
setup_environment_config(path_to_config)
}
\arguments{
\item{path_to_config}{String representing a path to a valid configuration file.}
}
\value{
list - A configuration list for the further use in the program
}
\description{
This functions loads a given configuration file, validates it and sets up
the logging capabilities of the R package {logger}.
}
