% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_crossover.R
\name{shuffled_crossover}
\alias{shuffled_crossover}
\title{Shuffled Crossover}
\usage{
shuffled_crossover(parent1, parent2, seed = NULL)
}
\arguments{
\item{parent1}{vec\if{html}{\out{<Some>}} - Vector containing the genome information of parent 1}

\item{parent2}{vec\if{html}{\out{<Some>}} - Vector containing the genome information of parent 2}

\item{seed}{Optional<NULL|integer> - A random seed for deterministic outcomes}
}
\value{
Optional<error|list<vec\if{html}{\out{<bool>}},vec\if{html}{\out{<bool>}}>>
}
\description{
Perform a shuffled crossover, e.g. shuffle all genes from both parents repectively and perform a single point
crossover to return two new children. Accepts an optional seed for deterministic outcomes.
}
