% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_mutation.R
\name{swap_mutation}
\alias{swap_mutation}
\title{Swap Mutation}
\usage{
swap_mutation(individual, seed = NULL)
}
\arguments{
\item{individual}{vec\if{html}{\out{<bool>}} - A boolean vector representating the genome information}

\item{seed}{Optional<NULL|integer> - A random seed for deterministic outcomes}
}
\value{
Optional<err|vec\if{html}{\out{<bool>}}>
}
\description{
This mutation algorithm generates two indices randomly, than swaps all values between the two generated indices.
The function can also take in an optional seed value for deterministic results.
}
