% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_selection.R
\name{tournament_selection}
\alias{tournament_selection}
\title{Tournament selection}
\usage{
tournament_selection(fitness_values, num_parents, tournament_size, seed = NULL)
}
\arguments{
\item{fitness_values}{Vec\if{html}{\out{<numeric>}} - A vector of numeric values representing the individuals fitness values}

\item{num_parents}{integer - The number of individuals to choose}

\item{tournament_size}{integer - The size of each tournament}

\item{seed}{Optional<NULL|integer> - A random seed for deterministic outcomes}
}
\value{
vec\if{html}{\out{<integer>}}
}
\description{
Perform \code{num_parents} tournamens of size \code{tournament_size}. For each tournament, a random selection
of size \code{torunament_size} individuals are chosen. The fittest individual within the tournament is added to the vector of chosen individuals.
}
