% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ga_crossover.R
\name{uniform_crossover}
\alias{uniform_crossover}
\title{Uniform Crossover}
\usage{
uniform_crossover(parent1, parent2, probability = 0.5, seed = NULL)
}
\arguments{
\item{parent1}{vec\if{html}{\out{<Some>}} - Vector containing the genome information of parent 1}

\item{parent2}{vec\if{html}{\out{<Some>}} - Vector containing the genome information of parent 2}

\item{seed}{Optional<NULL|integer> - A random seed for deterministic outcomes}

\item{proability}{numeric - Numerical value between 0 and 1}
}
\value{
Optional<error|list<vec\if{html}{\out{<bool>}},vec\if{html}{\out{<bool>}}>>
}
\description{
Performs a crossover between two individuals by determining on each gene a random point, comparing to a given
probability between 0 and 1. If the value is above the given probability, the a crossover happens,
else the gene remains stable. Accepts a seed to provide deterministic outcomes if needed.
}
