% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregator.R
\name{validate_aggregator_input}
\alias{validate_aggregator_input}
\title{Validate input arguments for the aggregator}
\usage{
validate_aggregator_input(
  index_vars = NULL,
  numerical_vars = NULL,
  agg_methods_num = NULL,
  categorical_vars = NULL,
  agg_methods_cat = NULL
)
}
\arguments{
\item{index_vars}{vec\if{html}{\out{<character>}} - A column name or vector of column names present in \code{data} to aggregate by}

\item{numerical_vars}{vec\if{html}{\out{<character>}} - A vector of column names for numerical aggregations}

\item{agg_methods_num}{vec\if{html}{\out{<character>}} - A vector of aggregation methods for each specified column in \code{numerical_vars}}

\item{categorical_vars}{vec\if{html}{\out{<character>}} - A vector of column names for categorical aggregations}

\item{agg_methods_categorical}{vec\if{html}{\out{<character>}} - A vector of aggregation methods for each specified column in \code{categorical_vars}}
}
\description{
This function performs a small set of sanitizer tasks to check if the provided input is
viable for the aggregator to run.
}
